Add tes3ui.getViewportSize(). Fix types on CameraData width/height.

+++ MWSE/TES3WorldController.h
@@ -35,8 +35,8 @@ namespace TES3 {
 			float fovDegrees; // 0x8
 			float nearPlaneDistance; // 0xC
 			float farPlaneDistance; // 0x10
-			float viewportWidth; // 0x14
-			float viewportHeight; // 0x18
+			unsigned int viewportWidth; // 0x14
+			unsigned int viewportHeight; // 0x18
 
 			//
 			// Other related this-call functions.
